# perf: improve performance of parentPaths