--- a/lib/helpers/path/parentPaths.js
+++ b/lib/helpers/path/parentPaths.js
@@ -1,13 +1,16 @@
 'use strict';
 
 module.exports = function parentPaths(path) {
-  const pieces = path.split('.');
-  let cur = '';
-  const ret = [];
-  for (let i = 0; i < pieces.length; ++i) {
-    cur += (cur.length > 0 ? '.' : '') + pieces[i];
-    ret.push(cur);
+  if (path.indexOf('.') === -1) {
+    return [path];
   }
-
+  let pos = path.indexOf('.')
+  const ret = [];
+  let i = 0;
+  while (pos !== -1) {
+    ret[i++] = (path.substring(0, pos));
+    pos = path.indexOf('.', pos + 1);
+  }
+  ret[i] = path.substring(0);
   return ret;
 };