--- a/lib/helpers/path/parentPaths.js
+++ b/lib/helpers/path/parentPaths.js
@@ -4,7 +4,7 @@ module.exports = function parentPaths(path) {
   if (path.indexOf('.') === -1) {
     return [path];
   }
-  let pos = path.indexOf('.')
+  let pos = path.indexOf('.');
   const ret = [];
   let i = 0;
   while (pos !== -1) {