--- a/lib/helpers/path/parentPaths.js
+++ b/lib/helpers/path/parentPaths.js
@@ -1,16 +1,18 @@
 'use strict';
 
+const dotRE = /\./g;
 module.exports = function parentPaths(path) {
   if (path.indexOf('.') === -1) {
     return [path];
   }
-  let pos = path.indexOf('.');
-  const ret = [];
-  let i = 0;
-  while (pos !== -1) {
-    ret[i++] = (path.substring(0, pos));
-    pos = path.indexOf('.', pos + 1);
+  const pieces = path.split(dotRE);
+  const len = pieces.length;
+  const ret = new Array(len);
+  let cur = '';
+  for (let i = 0; i < len; ++i) {
+    cur += (cur.length !== 0) ? '.' + pieces[i] : pieces[i];
+    ret[i] = cur;
   }
-  ret[i] = path.substring(0);
+
   return ret;
 };