--- a/test/helpers/parentPaths.test.js
+++ b/test/helpers/parentPaths.test.js
@@ -3,14 +3,14 @@
 const assert = require('assert');
 const parentPaths = require('../../lib/helpers/path/parentPaths');
 
-describe('parentPaths', function () {
-  it('first', function () {
+describe('parentPaths', function() {
+  it('first', function() {
     assert.deepEqual(parentPaths('first'), ['first']);
   });
-  it('first.second', function () {
+  it('first.second', function() {
     assert.deepEqual(parentPaths('first.second'), ['first', 'first.second']);
   });
-  it('first.', function () {
+  it('first.', function() {
     assert.deepEqual(parentPaths('first.'), ['first', 'first.']);
   });
 });