@Uzlopak 2022-01-24
I think now it is more stable in performance. You could only improve perf by caching...