@vkarpov15 2022-01-24
Based on your original comment, it looks like with these changes the perf degrades relative to the current code if you increase the number of `.` in the string. Is that still the case after your more recent changes?