@Uzlopak 2022-01-24
The original Code was giving false Benchmarks. More optimistic than it was actually. So I decided to make it more conservative code change and don't try to outsmart V8.